+++ src/lib/modelling/inference.py
@@ -4,7 +4,7 @@ from typing import Any, List, Dict
 
 from sklearn.preprocessing import OrdinalEncoder
 
-from src.lib.processing import CSVProcessingHandler
+from src.lib.data_managers import CSVProcessingHandler
 from src.lib.constants.columns import INFERENCE_COLUMNS
